Refactor: talk 경로 제거
