+++ src/app/child/[childId]/free/page.tsx
@@ -1,6 +1,6 @@
 import { Fetcher } from '@/lib/fetcher';
 import { API } from '@/constants/api';
-import TalkSession from '../_components/TalkSession';
+import TalkSession from '../talk/_components/TalkSession';
 import { ChildPageParams } from '@/types/page-props';
 
 // 서버 컴포넌트로, 페이지에 진입하기 전에 데이터를 미리 가져옵니다.
@@ -28,7 +28,7 @@ export default async function FreeTalkPage({ params }: ChildPageParams) {
       return <div>대화를 시작하는 데 실패했습니다.</div>;
     }
   } catch (error) {
-    console.error('[/talk/free] Fetch Error:', error);
+    console.error('[/free] Fetch Error:', error);
     return <div>서버와 통신하는 중 오류가 발생했습니다.</div>;
   }
 }
